servers/youtube-to-mp3/requirements.txt
mcp>=1.0.0
yt-dlp>=2024.0.0
mutagen>=1.47.0
Pillow>=10.0.0
requests>=2.31.0
fastapi>=0.104.0
uvicorn>=0.24.0
google-api-python-client>=2.100.0
google-auth>=2.23.0
google-auth-oauthlib>=1.1.0
google-auth-httplib2>=0.1.1
